camera: fix selinux=enforcing
the ultra hard to find selinux policies needed to fix the selinux
enforcing mode since Android 10 on the G4.

this also switches the user used to start the main cam services from
"cameraserver" to "camera":
- android.hardware.camera.provider@2.4-service.rc
- android.hardware.camera.provider@2.4-external-service.rc

this was required as the mm-qcamera-daemon creates sockets with the "camera" as user
and attempts to fix this had no success. instead all 3 main services
are now running with the "camera" user:

> u:r:hal_camera_default:s0      camera          589      1   22120   5604 binder_ioctl f2776954 S android.hardware.camera.provider@2.4-external-service
> u:r:hal_camera_default:s0      camera          778      1  103212  19224 binder_ioctl f543b954 S android.hardware.camera.provider@2.4-service
> u:r:mm-qcamerad:s0             camera         1095      1  127376  32240 poll_schedule_timeout efa3eb14 S mm-qcamera-daemon

this required to also set proper permissions on the devices needed for the camera in
- init.qcom.rc

all selinux policies added in hal_camera_default.te have a description
why they are needed and any single one is required as it is while it was ensured to just
set the minimal set required.
as there were no log denials for any of them it was REALLY hard identifying these.
read here how: https://gist.github.com/steadfasterX/e674749d6ca475ca8f4facdb7244e964#ultra-advanced-debug

Signed-off-by: steadfasterX <steadfasterX@gmail.com>